backend/backend/requirements.txt
asgiref==3.3.0
certifi==2020.6.20
cloudinary==1.22.0
dj-database-url==0.5.0
Django==3.0.9
django-cors-headers==3.5.0
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
numpy==1.19.2
pandas==1.1.3
psycopg2==2.8.6
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.10
whitenoise==5.2.0
